Breaking Change in concept, from right inclusive to left inclusive, for starts_ends
Earlier, a label was inferred to be active in range (start, end], with the understanding that we don't know what happened at start
because we don't know what happened (1/samplerate) before it.

To make it more in line with normal zero-based array indexing, and 24-hr clocks, the behavior is now changed to have labels
active in the range [start, end), with the understanding that the label for a time-point is the one that is active starting at said
time-point and ending in the next (1/samplerate).

Spectrograms time-points, that are calculated to be the center of the FFT windows, will now get labels for the samples from the right half of their window
rather than the left half. It's still wrong but... oh well...

The change in code was surprisingly minor and contained. Not for what's coming up though!