~ Fix failing pylint on old scripts
The old scripts have been kept for record-keeping reasons.
These scripts are not directly runnable with the current set of
requirements in the setup.py file (in particular, the scripts require
Keras, which is not one of the required dependencies).

The culprit directories are being ignored in the main pylintrc file of
the project. It is not the best solution, since the problem is with the
examples directory. But there was no quick and straightforward solution
right now, and the basename of the directory is _relatively_ unique.